+++ app/components/global/parts/WordAccordion.tsx
