+++ app/features/cards/components/TextPairManager.tsx
