+++ app/features/quiz/components/ImageModal.tsx
