+++ app/features/quiz/components/Result.tsx
