+++ app/features/slide/types/index.ts
@@ -1,0 +1,6 @@
+export interface SlideWord {
+	id: string;
+	word?: string;
+	translation?: string;
+	img?: string;
+}
