+++ app/types/word.ts
@@ -1,10 +1,3 @@
 export interface WordProps {
 	word: string;
 }
-
-export interface SlideWord {
-	id: string;
-	word?: string;
-	translation?: string;
-	img?: string;
-}
